requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via geoip2
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
dnspython==2.6.1
    # via
    #   -r requirements.in
    #   pymongo
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
geoip2==4.8.0
    # via -r requirements.in
idna==3.8
    # via
    #   requests
    #   yarl
maxminddb==2.6.2
    # via geoip2
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
pymongo==4.9.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   geoip2
scapy==2.6.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via multidict
urllib3==2.2.3
    # via requests
yarl==1.11.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
